requirements.txt
asgiref==3.11.1
attrs==26.1.0
cloudinary==1.44.2
colorama==0.4.6
coverage==7.13.5
dj-database-url==2.3.0
Django==6.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.2
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.20.0
PyJWT==2.12.1
pytest==8.4.2
pytest-django==4.12.0
python-decouple==3.8
PyYAML==6.0.3
requests==2.33.1
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.12
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
uritemplate==4.2.0
whitenoise==6.12.0
